+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixed
 * --check should fail if the number of newlines is different. [#2461](https://github.com/fsprojects/fantomas/issues/2461)
+* Comment after closing brace of computation expression is lost. [#2466](https://github.com/fsprojects/fantomas/issues/2466)
 
 ## [5.0.0-beta-008] - 2022-08-30
 
@@ -14,7 +15,7 @@
 * Idempotency problem when using a list with interpolated strings. [#2242](https://github.com/fsprojects/fantomas/issues/2242)
 * Idempotency problem when using set and spread syntax. [#2392](https://github.com/fsprojects/fantomas/issues/2392)
 * Formatting (*) returns ( * ). [#2452](https://github.com/fsprojects/fantomas/issues/2452)
-* Comment after last case of function with operator is removed.[#2454](https://github.com/fsprojects/fantomas/issues/2454)
+* Comment after last case of function with operator is removed. [#2454](https://github.com/fsprojects/fantomas/issues/2454)
 * Trailing whitespace in block comments should be preserved. [#2450](https://github.com/fsprojects/fantomas/issues/2450)
 
 ## [5.0.0-beta-007] - 2022-08-19